pyproject.toml
[tool.poetry]
name = "schedule-appointments"
version = "0.1.0"
description = ""
authors = ["Afroze Khan <itsmeafroze@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
pytest-django = "^4.1.0"
factory-boy = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
